test(agent): keep cross-project regression bounded
