+++ framework/core/tests/python/test_assignment_profile_source.py
@@ -1,12 +1,17 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import importlib
+import json
 from types import SimpleNamespace
 
 from kungfu import assignment_orchestration
+from kungfu.agent import run_agent
+from kungfu.workspace import resolve_workspace_target
+from agent_bootstrap_fixtures import verified_bootstrap_receipt
 
 
 ASSIGNMENT_CLI = importlib.import_module("kungfu.cli.commands.assignment")
+ROOT_HASH = "sha256:" + "a" * 64
 
 
 def _sha256(marker):
@@ -36,7 +41,7 @@ def test_kickoff_restores_work_control_after_dogfood_profile(monkeypatch, tmp_pa
     monkeypatch.setattr(
         ASSIGNMENT_CLI.run_agent,
         "bind_current_native_work",
-        lambda *_args: None,
+        lambda *_args, **_kwargs: None,
     )
     monkeypatch.setattr(
         ASSIGNMENT_CLI,
@@ -89,3 +94,84 @@ def test_assignment_profile_source_prefers_native_source_layout(tmp_path, monkey
     )
 
     assert ASSIGNMENT_CLI._profile_source() == native
+
+
+def test_claim_binds_work_from_explicit_external_project(monkeypatch, tmp_path):
+    requests = []
+    observed_runtime_dirs = []
+    console_project = tmp_path / "console-project"
+    console_runtime = console_project / ".kungfu" / "runtime"
+    console_runtime.mkdir(parents=True)
+    console_target = resolve_workspace_target(
+        "read-only", str(console_project), cwd=str(console_project)
+    )
+    work_project = tmp_path / "work-project"
+    work_runtime = work_project / ".kungfu" / "runtime"
+    work_runtime.mkdir(parents=True)
+    work_target = resolve_workspace_target(
+        "read-only", str(work_project), cwd=str(work_project)
+    )
+    envelope = {
+        "workspaceId": console_target.identity.workspace_id,
+        "consoleId": f"assistant:{console_target.identity.workspace_id}:native:one",
+        "attemptId": "native:one",
+    }
+    monkeypatch.setenv("KUNGFU_AGENT_CONSOLE_ENVELOPE", json.dumps(envelope))
+    monkeypatch.setenv(
+        "KUNGFU_AGENT_BOOTSTRAP_RECEIPT",
+        json.dumps(verified_bootstrap_receipt()),
+    )
+    monkeypatch.setenv("KUNGFU_AGENT_ATTEMPT_ID", "native:one")
+    monkeypatch.setenv("KUNGFU_AGENT_SESSION_ACTOR", "native:codex:native:one")
+    monkeypatch.setenv("KUNGFU_WORKSPACE_ROOT", str(console_project))
+    monkeypatch.setenv("KUNGFU_AGENT_RUNTIME_DIR", str(console_runtime))
+
+    def status(runtime_dir, *_args):
+        observed_runtime_dirs.append(runtime_dir)
+        return {
+            "assignment": {"assignment_id": "assignment:external"},
+            "query_proof_root": ROOT_HASH,
+        }
+
+    monkeypatch.setattr(ASSIGNMENT_CLI, "_status", status)
+    monkeypatch.setattr(ASSIGNMENT_CLI, "_profile_source", lambda: tmp_path)
+
+    def validate_source(_source, runtime_dir):
+        observed_runtime_dirs.append(runtime_dir)
+        return {
+            "inspection": {
+                "profile": {"id": "kungfu.work-control"},
+                "profile_suite_root": ROOT_HASH,
+            }
+        }
+
+    monkeypatch.setattr("kungfu.profile_sdk.validate_source", validate_source)
+
+    def invoke(request, **_kwargs):
+        requests.append(request)
+        if request["operation"] == "plan-native-bind-work":
+            return {
+                "operation": "native-bind-work",
+                "root": ROOT_HASH,
+                **request["input"]["session"],
+                "workRef": request["input"]["workRef"],
+            }
+        return {"status": "bound", "receiptRoot": ROOT_HASH}
+
+    monkeypatch.setattr(run_agent.session_surface, "invoke", invoke)
+    result = run_agent.bind_current_native_work(
+        str(work_runtime),
+        "initiative:external",
+        "assignment:external",
+        work_workspace_root=str(work_project),
+    )
+
+    assert observed_runtime_dirs == [str(work_runtime), str(work_runtime)]
+    assert result["workRef"]["workspaceId"] == work_target.identity.workspace_id
+    assert result["session"]["workConsoleId"] == envelope["consoleId"]
+    assert requests[0]["input"]["bindingScope"] == "explicit-external-project"
+    assert requests[0]["input"]["sourceWorkspaceId"] == envelope["workspaceId"]
+    assert [request["operation"] for request in requests] == [
+        "plan-native-bind-work",
+        "bind-native-work",
+    ]
